fix: Refactor a bit
