+++ package/android/src/main/java/com/mrousavy/camera/react/CameraView+Events.kt
@@ -10,8 +10,8 @@ import com.google.mlkit.vision.barcode.common.Barcode
 import com.mrousavy.camera.core.CameraError
 import com.mrousavy.camera.core.CodeScannerFrame
 import com.mrousavy.camera.core.UnknownCameraError
-import com.mrousavy.camera.core.code
-import com.mrousavy.camera.core.types.*
+import com.mrousavy.camera.core.types.CodeType
+import com.mrousavy.camera.core.types.ShutterType
 
 fun CameraView.invokeOnInitialized() {
   Log.i(CameraView.TAG, "invokeOnInitialized()")
